feat: make CategoryCard collapsible, reorder searched category to the top, destructure CategoryCard and add ShelterService as a new component
